added entity models for ledger management

+++ Features/Suppliers/Command/CreateSuppliers/CreateSupplierCommand.cs
@@ -11,6 +11,6 @@ namespace Inventory_Management_System.Features.Suppliers.Command.CreateSuppliers
         public string PhoneNumber { get; set; }
         public string Email { get; set; }
         public string NID { get; set; }
-        public int OpeningBalance { get; set; }
+        public decimal OpeningBalance { get; set; }
     }
 }
